+++ src/main/java/janggi/domain/piece/generator/PieceGenerator.java
@@ -1,9 +1,9 @@
-package janggi.domain.piece.gererator;
+package janggi.domain.piece.generator;
 
 import janggi.domain.piece.Piece;
 import java.util.List;
 
-public interface ChoPieceGenerator {
+public interface PieceGenerator {
 
     List<Piece> generate(KnightElephantSetting knightElephantSetting);
 }
