+++ src/test/java/janggi/domain/board/JanggiBoardTest.java
@@ -4,11 +4,11 @@ import janggi.domain.Position;
 import janggi.domain.Side;
 import janggi.domain.piece.Elephant;
 import janggi.domain.piece.Rook;
-import janggi.domain.piece.gererator.ChoPieceGenerator;
-import janggi.domain.piece.gererator.DefaultChoPieceGenerator;
-import janggi.domain.piece.gererator.DefaultHanPieceGenerator;
-import janggi.domain.piece.gererator.HanPieceGenerator;
-import janggi.domain.piece.gererator.KnightElephantSetting;
+import janggi.domain.piece.generator.ChoPieceGenerator;
+import janggi.domain.piece.generator.DefaultChoPieceGenerator;
+import janggi.domain.piece.generator.DefaultHanPieceGenerator;
+import janggi.domain.piece.generator.HanPieceGenerator;
+import janggi.domain.piece.generator.KnightElephantSetting;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
